+++ app/Http/Controllers/Admin/WelcomeSectionController.php
@@ -35,7 +35,7 @@ class WelcomeSectionController extends Controller
             $leftFileName = time() . '-' . uniqid() . '.' . $file->getClientOriginalExtension();
             $file->move(public_path('uploads/welcome-sections'), $leftFileName);
             
-            if ($welcomeSection->left_image && file_exists(public_path('uploads/welcome-sections/' . $welcomeSection->left_image))) {
+            if ($welcomeSection->left_image && $welcomeSection->left_image !== 'home-welcome-left.jpg' && file_exists(public_path('uploads/welcome-sections/' . $welcomeSection->left_image))) {
                 unlink(public_path('uploads/welcome-sections/' . $welcomeSection->left_image));
             }
         }
@@ -47,7 +47,7 @@ class WelcomeSectionController extends Controller
             $rightFileName = time() . '-' . uniqid() . '.' . $file->getClientOriginalExtension();
             $file->move(public_path('uploads/welcome-sections'), $rightFileName);
             
-            if ($welcomeSection->right_image && file_exists(public_path('uploads/welcome-sections/' . $welcomeSection->right_image))) {
+            if ($welcomeSection->right_image && $welcomeSection->right_image !== 'home-welcome-right.jpg' && file_exists(public_path('uploads/welcome-sections/' . $welcomeSection->right_image))) {
                 unlink(public_path('uploads/welcome-sections/' . $welcomeSection->right_image));
             }
         }
